test_app/Drain3/requirements.txt
cachetools==4.2.1
certifi==2024.2.2
click==8.1.7
clickhouse-connect==0.7.8
distlib==0.3.8
drain3==0.9.11
filelock==3.13.4
joblib==1.4.2
jsonpickle==1.5.1
kafka-python==2.0.2
lz4==4.3.3
nltk==3.8.1
numpy==1.26.4
pandas==2.2.1
pipenv==2023.12.1
platformdirs==4.2.1
python-dateutil==2.9.0.post0
pytz==2024.1
regex==2022.3.2
scipy==1.13.0
six==1.16.0
tqdm==4.66.2
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.26.0
watchdog==4.0.0
zstandard==0.22.0
